fix null in active cluster file

+++ shared/configuration.ts
@@ -62,7 +62,9 @@ export class Configuration {
             clients: [...config.clients, ...this.clients.map(c => c.name)].filter((c, i, a) => a.indexOf(c) == i)
         }));
 
-        writeFileSync(Cluster.activeClusterFileLocation, JSON.stringify(this.activeCluster || null));
+        if (this.activeCluster || !existsSync(Cluster.activeClusterFileLocation)) {
+            writeFileSync(Cluster.activeClusterFileLocation, JSON.stringify(this.activeCluster || null));
+        }
     }
 
     static load() {
